Added an Appointment Cancellation module to the Web API project
