+++ src/Extensions/ServiceCollectionExtensions.cs
@@ -20,6 +20,7 @@ public static class ServiceCollectionExtensions
                 .AddScoped<IRoleService, RoleService>()
                 .AddScoped<IOfficeService, OfficeService>()
                 .AddScoped<IAppointmentService, AppointmentService>()
+                .AddScoped<IAppointmentCancellationService, AppointmentCancellationService>()
                 .AddScoped<IEmployeeScheduleService, EmployeeScheduleService>()
                 .AddScoped<IFavoriteDentistService, FavoriteDentistService>()
                 .AddScoped<IOfficeScheduleService, OfficeScheduleService>()
@@ -44,6 +45,7 @@ public static class ServiceCollectionExtensions
                 .AddScoped<IOfficeRepository, OfficeRepository>()
                 .AddScoped<ISpecificTreatmentRepository, SpecificTreatmentRepository>()
                 .AddScoped<IAppointmentRepository, AppointmentRepository>()
+                .AddScoped<IAppointmentCancellationRepository, AppointmentCancellationRepository>()
                 .AddScoped<IEmployeeScheduleRepository, EmployeeScheduleRepository>()
                 .AddScoped<IFavoriteDentistRepository, FavoriteDentistRepository>()
                 .AddScoped<IOfficeScheduleRepository, OfficeScheduleRepository>()
