+++ src/Features/AppointmentCancellation/DTOs/AppointmentCancelDto.cs
@@ -1,0 +1,7 @@
+﻿namespace DentallApp.Features.AppointmentCancellation.DTOs;
+
+public class AppointmentCancelDto
+{
+    public string Reason { get; set; }
+    public IEnumerable<AppointmentCancelDetailsDto> Appointments { get; set; }
+}
