+++ src/Features/AppointmentCancellation/DTOs/AppointmentsThatCannotBeCanceledDto.cs
@@ -1,0 +1,6 @@
+﻿namespace DentallApp.Features.AppointmentCancellation.DTOs;
+
+public class AppointmentsThatCannotBeCanceledDto
+{
+    public IEnumerable<int> AppointmentsId { get; set; }
+}
