+++ src/Features/AppointmentCancellation/IAppointmentCancellationService.cs
@@ -1,0 +1,7 @@
+﻿namespace DentallApp.Features.AppointmentCancellation;
+
+public interface IAppointmentCancellationService
+{
+    Task<Response> CancelBasicUserAppointmentAsync(int appointmentId, int currentUserId);
+    Task<Response<AppointmentsThatCannotBeCanceledDto>> CancelAppointmentsAsync(ClaimsPrincipal currentEmployee, AppointmentCancelDto appointmentCancelDto);
+}
