+++ src/Features/Appointments/IAppointmentRepository.cs
@@ -16,18 +16,4 @@ public interface IAppointmentRepository : IRepository<Appointment>
     /// </summary>
     /// <returns><c>true</c> sí la fecha y hora de la cita no está disponible, de lo contrario devuelve <c>false</c>.</returns>
     Task<bool> IsNotAvailableAsync(AppointmentInsertDto appointmentDto);
-
-    /// <summary>
-    /// Cancela una o más citas agendadas de un consultorio.
-    /// </summary>
-    /// <param name="officeId">El ID del consultorio.</param>
-    /// <param name="appointmentsId">Un conjunto de ID de citas a cancelar.</param>
-    Task<int> CancelAppointmentsByOfficeIdAsync(int officeId, IEnumerable<int> appointmentsId);
-
-    /// <summary>
-    /// Cancela una o más citas agendadas de un dentista.
-    /// </summary>
-    /// <param name="dentistId">El ID del dentista.</param>
-    /// <param name="appointmentsId">Un conjunto de ID de citas a cancelar.</param>
-    Task<int> CancelAppointmentsByDentistIdAsync(int dentistId, IEnumerable<int> appointmentsId);
 }
