+++ src/Features/Appointments/IAppointmentService.cs
@@ -2,9 +2,7 @@
 
 public interface IAppointmentService
 {
-    Task<Response> CancelBasicUserAppointmentAsync(int appointmentId, int currentUserId);
     Task<Response> CreateAppointmentAsync(AppointmentInsertDto appointmentInsertDto);
     Task<Response> UpdateAppointmentAsync(int id, ClaimsPrincipal currentEmployee, AppointmentUpdateDto appointmentUpdateDto);
-    Task<Response<AppointmentsThatCannotBeCanceledDto>> CancelAppointmentsAsync(ClaimsPrincipal currentEmployee, AppointmentCancelDto appointmentCancelDto);
     Task<Response<IEnumerable<AppointmentGetByEmployeeDto>>> GetAppointmentsForEmployeeAsync(ClaimsPrincipal currentEmployee, AppointmentPostDateDto appointmentPostDto);
 }
